+++ examples/vqeh2o_benchmark.py
@@ -1,6 +1,11 @@
 """
 Time comparison for different evaluation approach on molecule VQE
 """
+
+import os
+
+os.environ["TF_FORCE_GPU_ALLOW_GROWTH"] = "true"
+# one need this for jax+gpu combination in some cases
 import time
 import numpy as np
 from openfermion.chem import MolecularData
@@ -36,7 +41,7 @@ print("%s terms in H2O qubit Hamiltonian" % len(wb))
 mb = tc.quantum.PauliStringSum2COO_numpy(lsb, wb)
 mbd = mb.todense()
 mb = K.coo_sparse_matrix(
-    np.transpose(np.stack([mb.row, mb.col])), mb.data, shape=(2 ** n, 2 ** n)
+    np.transpose(np.stack([mb.row, mb.col])), mb.data, shape=(2**n, 2**n)
 )
 mbd = tc.array_to_tensor(mbd)