+++ tensorcircuit/circuit.py
@@ -13,7 +13,7 @@ import tensornetwork as tn
 
 from . import gates
 from .cons import backend, contractor, dtypestr, npdtype
-from .quantum import QuVector, QuOperator
+from .quantum import QuVector, QuOperator, identity
 from .simplify import _split_two_qubit_gate
 from .vis import qir2tex
 
@@ -59,7 +59,12 @@ class Circuit:
     ]
     mpogates = ["multicontrol", "mpo"]
 
-    gate_alias_list = [["cnot", "cx"], ["fredkin", "cswap"], ["toffoli", "ccnot"]]
+    gate_alias_list = [
+        ["cnot", "cx"],
+        ["fredkin", "cswap"],
+        ["toffoli", "ccnot"],
+        ["any", "unitary"],
+    ]
 
     def __init__(
         self,
@@ -199,7 +204,10 @@ class Circuit:
                             other = (e.node2, e.axis2)
                         e.disconnect()
                         new_front[j] ^ other[0][other[1]]
+        j += 1
+        self._front += new_front[j:]
         self._nodes = new_nodes + self._nodes[self._start_index :]
+        self._start_index = len(new_nodes)
 
     @classmethod
     def _meta_apply(cls) -> None:
@@ -549,6 +557,13 @@ class Circuit:
         return apply
 
     def get_quvector(self) -> QuVector:
+        """
+        Get the representation of the output state in the form of ``QuVector``
+        while maintaining the circuit uncomputed
+
+        :return: ``QuVector`` representation of the output state from the circuit
+        :rtype: QuVector
+        """
         _, edges = self._copy()
         return QuVector(edges)
 
@@ -592,7 +607,6 @@ class Circuit:
         """
         return self._qir
 
-    # TODO(@refraction-ray): derive visualization and serialization from the consistent qir
     # TODO(@refraction-ray): add noise support in IR
     # TODO(@refraction-ray): IR for density matrix simulator?
 
@@ -683,6 +697,7 @@ class Circuit:
         """
         Middle measurement in z-basis on the circuit, note the wavefunction output is not normalized
         with ``mid_measurement`` involved, one should normalize the state manually if needed.
+        This is a post-selection method as keep is provided as a prior.
 
         :param index: The index of qubit that the Z direction postselection applied on.
         :type index: int
@@ -715,6 +730,45 @@ class Circuit:
         self._front[index] = mg2.get_edge(0)
         self._nodes.append(mg1)
         self._nodes.append(mg2)
+
+    mid_measure = mid_measurement
+
+    def cond_measurement(self, index: int) -> Tensor:
+        """
+        Measurement on z basis at ``index`` qubit based on quantum amplitude
+        (not post-selection). The highlight is that this method can return the
+        measured result as a int Tensor and thus maintained a jittable pipeline.
+
+        :Example:
+
+        >>> c = tc.Circuit(2)
+        >>> c.H(0)
+        >>> r = c.cond_measurement(0)
+        >>> c.conditional_gate(r, [tc.gates.i(), tc.gates.x()], 1)
+        >>> c.expectation([tc.gates.z(), [0]]), c.expectation([tc.gates.z(), [1]])
+        # two possible outputs: (1, 1) or (-1, -1)
+
+        :param index: the qubit for the z-basis measurement
+        :type index: int
+        :return: 0 or 1 for z measurement on up and down freedom
+        :rtype: Tensor
+        """
+        return self.general_kraus(
+            [np.array([[1.0, 0], [0, 0]]), np.array([[0, 0], [0, 1]])], index  # type: ignore
+        )
+
+    cond_measure = cond_measurement
+
+    def prepend(self, c: "Circuit") -> "Circuit":
+        self.replace_mps_inputs(c.quvector())
+        self._qir = c._qir + self._qir
+        return self
+
+    def append(self, c: "Circuit") -> "Circuit":
+        c.replace_mps_inputs(self.quvector())
+        c._qir = self._qir + c._qir
+        self.__dict__ = c.__dict__
+        return self
 
     def depolarizing2(
         self,
@@ -758,7 +812,27 @@ class Circuit:
         py: float,
         pz: float,
         status: Optional[float] = None,
-    ) -> float:
+    ) -> Tensor:
+        """
+        Apply depolarizing channel in a Monte Carlo way,
+        i.e. for each call of this method, one of gates from
+        X, Y, Z, I are applied on the circuit based on the probability
+        indicated by ``px``, ``py``, ``pz``.
+
+        :param index: The qubit that depolarizing channel is on
+        :type index: int
+        :param px: probability for X noise
+        :type px: float
+        :param py: probability for Y noise
+        :type py: float
+        :param pz: probability for Z noise
+        :type pz: float
+        :param status: random seed uniformly from 0 to 1, defaults to None (generated implicitly)
+        :type status: Optional[float], optional
+        :return: int Tensor, the element lookup: [0: x, 1: y, 2: z, 3: I]
+        :rtype: Tensor
+        """
+
         # px/y/z here not support differentiation for now
         # jit compatible for now
         # assert px + py + pz < 1 and px >= 0 and py >= 0 and pz >= 0
@@ -771,21 +845,43 @@ class Circuit:
             )
             r = backend.cast(r / 2 + 1.5, dtype="int32")
             # [0: x, 1: y, 2: z, 3: I]
-            r = backend.onehot(r, 4)
-            r = backend.cast(r, dtype=dtypestr)
+
             return r
 
         if status is None:
             status = backend.implicit_randu()[0]
         r = step_function(status)
+        rv = backend.onehot(r, 4)
+        rv = backend.cast(rv, dtype=dtypestr)
         g = (
-            r[0] * gates._x_matrix
-            + r[1] * gates._y_matrix
-            + r[2] * gates._z_matrix
-            + r[3] * gates._i_matrix
+            rv[0] * gates._x_matrix
+            + rv[1] * gates._y_matrix
+            + rv[2] * gates._z_matrix
+            + rv[3] * gates._i_matrix
         )
         self.any(index, unitary=g)  # type: ignore
-        return 0.0
+        return r
+
+    def select_gate(self, which: Tensor, kraus: Sequence[Gate], *index: int) -> None:
+        """
+        Apply ``which``-th gate from ``kraus`` list, i.e. apply kraus[which]
+
+        :param which: Tensor of shape [] and dtype int
+        :type which: Tensor
+        :param kraus: A list of gate in the form of ``tc.gate`` or Tensor
+        :type kraus: Sequence[Gate]
+        :param index: the qubit lines the gate applied on
+        :type index: int
+        """
+        kraus = [k.tensor if isinstance(k, tn.Node) else k for k in kraus]
+        kraus = [gates.array_to_tensor(k) for k in kraus]
+        l = len(kraus)
+        r = backend.onehot(which, l)
+        r = backend.cast(r, dtype=dtypestr)
+        tensor = reduce(add, [r[i] * kraus[i] for i in range(l)])
+        self.any(*index, unitary=tensor)  # type: ignore
+
+    conditional_gate = select_gate
 
     def unitary_kraus2(
         self,
@@ -793,7 +889,7 @@ class Circuit:
         *index: int,
         prob: Optional[Sequence[float]] = None,
         status: Optional[float] = None,
-    ) -> float:
+    ) -> Tensor:
         # general impl from Monte Carlo trajectory depolarizing above
         # still jittable
         # speed is similar to ``unitary_kraus``
@@ -811,7 +907,19 @@ class Circuit:
         *index: int,
         prob: Optional[Sequence[float]] = None,
         status: Optional[float] = None,
-    ) -> float:
+    ) -> Tensor:
+        """
+        Apply unitary gates in ``kraus`` randomly based on corresponding ``prob``.
+
+        :param kraus: List of ``tc.gates.Gate`` or just Tensors
+        :type kraus: Sequence[Gate]
+        :param prob: prob list with the same size as ``kraus``, defaults to None
+        :type prob: Optional[Sequence[float]], optional
+        :param status: random seed between 0 to 1, defaults to None
+        :type status: Optional[float], optional
+        :return: shape [] int dtype tensor indicates which kraus gate is actually applied
+        :rtype: Tensor
+        """
         # general impl from Monte Carlo trajectory depolarizing above
         # still jittable
 
@@ -835,9 +943,10 @@ class Circuit:
         get_gate_from_index: Optional[
             Callable[[Tensor, Sequence[Tensor]], Tensor]
         ] = None,
-    ) -> float:  # DRY
+    ) -> Tensor:  # DRY
         sites = len(index)
         kraus = [k.tensor if isinstance(k, tn.Node) else k for k in kraus]
+        kraus = [gates.array_to_tensor(k) for k in kraus]
         if prob is None:
             prob = [
                 backend.real(backend.trace(backend.adjoint(k) @ k) / k.shape[0])
@@ -867,7 +976,7 @@ class Circuit:
         g = get_gate_from_index(r, kraus)
         g = backend.reshape(g, [2 for _ in range(sites * 2)])
         self.any(*index, unitary=g)  # type: ignore
-        return 0.0
+        return r
 
     def _general_kraus_tf(
         self,
@@ -942,7 +1051,7 @@ class Circuit:
         kraus: Sequence[Gate],
         *index: int,
         status: Optional[float] = None,
-    ) -> float:
+    ) -> Tensor:
         # the graph building time is frustratingly slow, several minutes
         # though running time is in terms of ms
         # raw running time in terms of s
@@ -954,7 +1063,7 @@ class Circuit:
         # layerwise jit technique can greatly boost the staging time, see in /examples/mcnoise_boost.py
         sites = len(index)
         kraus_tensor = [k.tensor if isinstance(k, tn.Node) else k for k in kraus]
-        # kraus_tensor = [k.tensor for k in kraus]
+        kraus_tensor = [gates.array_to_tensor(k) for k in kraus_tensor]
 
         # tn with hole
         newnodes, newfront = self._copy()
@@ -988,15 +1097,14 @@ class Circuit:
             for w, k in zip(prob, kraus_tensor)
         ]
 
-        self.unitary_kraus2(new_kraus, *index, prob=prob, status=status)
-        return 0.0
+        return self.unitary_kraus2(new_kraus, *index, prob=prob, status=status)
 
     def general_kraus(
         self,
         kraus: Sequence[Gate],
         *index: int,
         status: Optional[float] = None,
-    ) -> float:
+    ) -> Tensor:
         """
         Monte Carlo trajectory simulation of general Kraus channel whose Kraus operators cannot be
         amplified to unitary operators. For unitary operators composed Kraus channel, :py:meth:`unitary_kraus`
@@ -1010,8 +1118,8 @@ class Circuit:
         :type kraus: Sequence[Gate]
         :param index: The qubits index that Kraus channel is applied on.
         :type index: int
-        :param status: Random tensor between 0 or 1, defaults to be None,
-            the random number will be generated automatically
+        :param status: Random tensor uniformly between 0 or 1, defaults to be None,
+            when the random number will be generated automatically
         :type status: Optional[float], optional
         """
         return self._general_kraus_2(kraus, *index, status=status)
@@ -1057,6 +1165,7 @@ class Circuit:
         Compute the output wavefunction from the circuit.
 
         :param form: The str indicating the form of the output wavefunction.
+            "default": [-1], "ket": [-1, 1], "bra": [1, -1]
         :type form: str, optional
         :return: Tensor with the corresponding shape.
         :rtype: Tensor
@@ -1091,6 +1200,21 @@ class Circuit:
         return newnodes, newfront
 
     state = wavefunction
+
+    def matrix(self) -> Tensor:
+        """
+        Get the unitary matrix for the circuit irrespective with the circuit input state.
+
+        :return: The circuit unitary matrix
+        :rtype: Tensor
+        """
+        mps = identity([2 for _ in range(self._nqubits)])
+        c = Circuit(self._nqubits)
+        ns, es = self._copy()
+        c._nodes = ns
+        c._front = es
+        c.replace_mps_inputs(mps)
+        return backend.reshapem(c.state())
 
     def amplitude(self, l: str) -> tn.Node.tensor:
         """
@@ -1131,7 +1255,7 @@ class Circuit:
 
     def measure(self, *index: int, with_prob: bool = False) -> Tuple[str, float]:
         """
-        Take measurement to the given quantum lines.
+        Take measurement on the given quantum lines by ``index``.
 
         :Example:
 
@@ -1194,7 +1318,7 @@ class Circuit:
     ) -> Tuple[Tensor, Tensor]:
         """
         Take measurement to the given quantum lines.
-        This method implemented by jit is about 100 times faster than unjit version!
+        This method is jittable is and about 100 times faster than unjit version!
 
         :param index: Measure on which quantum line.
         :type index: int
@@ -1245,7 +1369,7 @@ class Circuit:
 
     def perfect_sampling(self) -> Tuple[str, float]:
         """
-        Sampling bistrings from the circuit output.
+        Sampling bistrings from the circuit output based on quantum amplitudes.
         Reference: arXiv:1201.3974.
 
         :return: Sampled bit string and the corresponding theoretical probability.
@@ -1260,6 +1384,14 @@ class Circuit:
     def expectation_before(
         self, *ops: Tuple[tn.Node, List[int]], reuse: bool = True
     ) -> List[tn.Node]:
+        """
+        Return the list of nodes that consititues the expectation value just before the contraction.
+
+        :param reuse: whether contract the output state firstly, defaults to True
+        :type reuse: bool, optional
+        :return: The tensor network for the expectation
+        :rtype: List[tn.Node]
+        """
         nodes1, edge1 = self._copy_state_tensor(reuse=reuse)
         nodes2, edge2 = self._copy_state_tensor(conj=True, reuse=reuse)
         nodes1.extend(nodes2)  # left right op order for plain contractor
@@ -1317,9 +1449,43 @@ class Circuit:
         nodes1 = self.expectation_before(*ops, reuse=reuse)
         return contractor(nodes1).tensor
 
+    def expectation_ps(
+        self,
+        x: Optional[Sequence[int]] = None,
+        y: Optional[Sequence[int]] = None,
+        z: Optional[Sequence[int]] = None,
+        reuse: bool = True,
+    ) -> Tensor:
+        """
+        Shortcut for Pauli string expectation.
+        x, y, z list are for X, Y, Z positions
+
+        :param x: _description_, defaults to None
+        :type x: Optional[Sequence[int]], optional
+        :param y: _description_, defaults to None
+        :type y: Optional[Sequence[int]], optional
+        :param z: _description_, defaults to None
+        :type z: Optional[Sequence[int]], optional
+        :param reuse: whether to cache and reuse the wavefunction, defaults to True
+        :type reuse: bool, optional
+        :return: Expectation value
+        :rtype: Tensor
+        """
+        obs = []
+        if x is not None:
+            for i in x:
+                obs.append([gates.x(), [i]])  # type: ignore
+        if y is not None:
+            for i in y:
+                obs.append([gates.y(), [i]])  # type: ignore
+        if z is not None:
+            for i in z:
+                obs.append([gates.z(), [i]])  # type: ignore
+        return self.expectation(*obs, reuse=reuse)  # type: ignore
+
     def to_qiskit(self) -> Any:
         """
-        Translate to a qiskit object.
+        Translate ``tc.Circuit`` to a qiskit QuantumCircuit object.
 
         :return: A qiskit object of this circuit.
         """
@@ -1329,7 +1495,6 @@ class Circuit:
         return qir2qiskit(qir, n=self._nqubits)
 
     def draw(self, **kws: Any) -> Any:
-        # TODO(@Xinghan): The example part of the documentation doesn't load as you put your mouse on the draw dunction
         """
         Visualise the circuit.
         This method recevies the keywords as same as qiskit.circuit.QuantumCircuit.draw.
@@ -1352,13 +1517,43 @@ class Circuit:
 
     @classmethod
     def from_qiskit(
-        self, qc: Any, n: int, inputs: Optional[List[float]] = None
+        self, qc: Any, n: Optional[int] = None, inputs: Optional[List[float]] = None
     ) -> "Circuit":
+        """
+        Import Qiskit QuantumCircuit object as a ``tc.Circuit`` object.
+
+        :Example:
+
+        >>> from qiskit import QuantumCircuit
+        >>> qisc = QuantumCircuit(3)
+        >>> qisc.h(2)
+        >>> qisc.cswap(1, 2, 0)
+        >>> qisc.swap(0, 1)
+        >>> c = tc.Circuit.from_qiskit(qisc)
+
+        :param qc: Qiskit Circuit object
+        :type qc: QuantumCircuit in Qiskit
+        :param n: The number of qubits for the circuit
+        :type n: int
+        :param inputs: possible input wavefunction for ``tc.Circuit``, defaults to None
+        :type inputs: Optional[List[float]], optional
+        :return: The same circuit but as tensorcircuit object
+        :rtype: Circuit
+        """
         from .translation import qiskit2tc
+
+        if n is None:
+            n = qc.num_qubits
 
         return qiskit2tc(qc.data, n, inputs)  # type: ignore
 
     def vis_tex(self, **kws: Any) -> str:
+        """
+        Generate latex string based on quantikz latex package
+
+        :return: Latex string that can be directly compiled via, e.g. latexit
+        :rtype: str
+        """
         return qir2tex(self._qir, self._nqubits, **kws)  # type: ignore
 
 
@@ -1373,7 +1568,7 @@ def to_graphviz(
 ) -> graphviz.Graph:
     """
     Not an ideal visualization for quantum circuit, but reserve here as a general approach to show the tensornetwork
-    [Deprecated, use ``tensorcircuit.vis.qir2tex instead``]
+    [Deprecated, use ``Circuit.vis_tex`` or ``Circuit.draw`` instead]
     """
     # Modified from tensornetwork codebase
     nodes = c._nodes